Fold the Picker demo into the shared inputs template
