+++ client-app/src/desktop/tabs/forms/index.ts
@@ -1,5 +1,4 @@
 export * from './FormPanel';
 export * from './DateRangePickerPanel';
-export * from './PickerPanel';
 export * from './ToolbarFormPanel';
 export * from './inputs';
