+++ client-app/src/desktop/tabs/forms/inputs/PickerPanel.scss
@@ -1,13 +1,4 @@
 .tb-picker-panel {
-  &__row {
-    margin-bottom: var(--xh-pad-double-px);
-    flex-shrink: 0;
-  }
-
-  &__label {
-    font-weight: 500;
-  }
-
   &__badge {
     margin-left: 6px;
     font-size: 11px;
